+++ js/jquery.terminal.js
@@ -4,7 +4,7 @@
  *  __ / // // // // // _  // _// // / / // _  // _//     // //  \/ // _ \/ /
  * /  / // // // // // ___// / / // / / // ___// / / / / // // /\  // // / /__
  * \___//____ \\___//____//_/ _\_  / /_//____//_/ /_/ /_//_//_/ /_/ \__\_\___/
- *           \/              /____/                              version 1.17.0
+ *           \/              /____/                              version DEV
  *
  * This file is part of jQuery Terminal. http://terminal.jcubic.pl
  *
@@ -32,7 +32,7 @@
  * Copyright (c) 2007-2013 Alexandru Marasteanu <hello at alexei dot ro>
  * licensed under 3 clause BSD license
  *
- * Date: Sun, 01 Jul 2018 13:35:08 +0000
+ * Date: Sun, 08 Jul 2018 07:08:23 +0000
  */
 
 /* TODO:
@@ -216,6 +216,28 @@
     }
 })(function($, wcwidth, undefined) {
     'use strict';
+    // -----------------------------------------------------------------------
+    // :: debug functions
+    // -----------------------------------------------------------------------
+    function debug(str) {
+        if (false) {
+            console.log(str);
+            $.terminal.active().echo(str);
+        }
+    }
+    /* commented out so it don't affect coverage
+    // -----------------------------------------------------------------------
+    Function.prototype.monitor = function() {
+        var name = this.name;
+        var fn = this;
+        return function(...args) {
+            console.log(name + '(' + JSON.stringify(args).replace(/^\[|\]$/g, '') + ')');
+            var ret = fn.apply(this, args);
+            console.log('return ' + JSON.stringify(ret));
+            return ret;
+        };
+    };
+    //*/
     // -----------------------------------------------------------------------
     // :: Replacemenet for jQuery 2 deferred objects
     // -----------------------------------------------------------------------
@@ -1655,13 +1677,13 @@
         // ---------------------------------------------------------------------
         function formatting(string) {
             // we don't want to format command when user type formatting in
-            string = $.terminal.escape_formatting(string);
             try {
+                string = $.terminal.escape_formatting(string);
                 string = $.terminal.apply_formatters(string, settings);
                 string = $.terminal.normalize(string);
                 return string;
             } catch (e) {
-                alert_exception('[Formatting]', e);
+                alert_exception('[Formatting]', e.stack);
                 return string;
             }
         }
@@ -1706,7 +1728,7 @@
             return result.join('');
         }
         // ---------------------------------------------------------------------
-        // :: shortcut helper
+        // :: shortcut helpers
         // ---------------------------------------------------------------------
         function length(str) {
             return $.terminal.length(str);
@@ -1734,69 +1756,30 @@
             }
             if (formatters.length) {
                 return formatters.reduce(function(result, frmt) {
+                    var options = frmt[2] || {};
+                    if (options.loop) {
+                        result = result.slice();
+                        while (result[0].match(frmt[0])) {
+                            result = $.terminal.tracking_replace(
+                                result[0],
+                                frmt[0],
+                                frmt[1],
+                                result[1]
+                            );
+                        }
+                        return result;
+                    }
                     var command = result[0];
                     var position = result[1];
-                    return tracking_replace(command, frmt[0], frmt[1], position);
+                    return $.terminal.tracking_replace(
+                        command,
+                        frmt[0],
+                        frmt[1],
+                        position
+                    );
                 }, [command, position])[1];
             }
             return position;
-        }
-        // ---------------------------------------------------------------------
-        // :: source https://stackoverflow.com/a/46756077/387194
-        // ---------------------------------------------------------------------
-        function tracking_replace(string, rex, replacement, position) {
-            var new_string = "";
-            var match;
-            var index = 0;
-            var rep_string;
-            var new_position = position;
-            var start;
-            rex.lastIndex = 0; // Just to be sure
-            while ((match = rex.exec(string))) {
-                // Add any of the original string we just skipped
-                var last_index = length(substring(string, 0, rex.lastIndex));
-                start = last_index - length(match[0]);
-                if (index < start) {
-                    new_string += substring(string, index, start);
-                }
-                index = last_index;
-                // Build the replacement string. This just handles $$ and $n,
-                // you may want to add handling for $`, $', and $&.
-                rep_string = replacement.replace(/\$(\$|\d)/g, function(m, c0) {
-                    if (c0 === "$") {
-                        return "$";
-                    }
-                    return match[c0];
-                });
-                // Add on the replacement
-                new_string += rep_string;
-                // If the position is affected...
-                if (start < position) {
-                    // ... update it:
-                    if (last_index < position) {
-                        // It's after the replacement, move it
-                        new_position = Math.max(
-                            0,
-                            new_position +
-                            length(rep_string) -
-                            length(match[0])
-                        );
-                    } else {
-                        // It's *in* the replacement, put it just after
-                        new_position += length(rep_string) - (position - start);
-                    }
-                }
-                // single replace
-                if (rex.flags.indexOf('g') === -1) {
-                    break;
-                }
-            }
-            // Add on any trailing text in the string
-            if (index < length(string)) {
-                new_string += substring(string, index);
-            }
-            // Return the string and the updated position
-            return [new_string, new_position];
         }
         // ---------------------------------------------------------------------
         // :: Function that displays the command line. Split long lines and
@@ -1867,7 +1850,7 @@
                         break;
                 }
                 var pos = formatted_position;
-                string = formatting(safe(string.replace(/&/g, '&amp;')));
+                string = formatting(string);
                 var i;
                 self.find('div:not(.cursor-line,.clipboard-wrapper)').remove();
                 before.html('');
@@ -2134,7 +2117,7 @@
                         string + command.slice(position);
                 }
                 if (!stay) {
-                    self.position(string.length, true);
+                    self.position(string.length, true, true);
                 } else {
                     fix_textarea();
                 }
@@ -2185,7 +2168,7 @@
             kill_text: function() {
                 return kill_text;
             },
-            position: function(n, relative) {
+            position: function(n, relative, silent) {
                 if (typeof n === 'number') {
                     if (relative) {
                         position += n;
@@ -2200,8 +2183,10 @@
                     if (is_function(settings.onPositionChange)) {
                         settings.onPositionChange(position);
                     }
-                    redraw();
-                    fix_textarea(true);
+                    if (!silent) {
+                        redraw();
+                        fix_textarea(true);
+                    }
                     return self;
                 } else {
                     return position;
@@ -2505,12 +2490,6 @@
             event.fake = true;
             doc.trigger(event);
         }
-        function debug(str) {
-            if (false) {
-                console.log(str);
-                $.terminal.active().echo(str);
-            }
-        }
         function input_event() {
             debug('input ' + no_keydown + ' || ' + process + ' ((' + no_keypress +
                   ' || ' + dead_key + ') && !' + skip_insert + ' && (' + single_key +
@@ -2732,10 +2711,11 @@
     }
     // -------------------------------------------------------------------------
     function safe(string) {
-        if (!string.match(/[<>]/)) {
+        if (!string.match(/[<>&]/)) {
             return string;
         }
-        return string.replace(/>/g, '&gt;').replace(/</g, '&lt;');
+        return string.replace(/&(?![^;]+;)/g, '&amp;')
+            .replace(/>/g, '&gt;').replace(/</g, '&lt;');
     }
     // -------------------------------------------------------------------------
     function crlf(string) {
@@ -2995,8 +2975,8 @@
     }
     // -------------------------------------------------------------------------
     $.terminal = {
-        version: '1.17.0',
-        date: 'Sun, 01 Jul 2018 13:35:08 +0000',
+        version: 'DEV',
+        date: 'Sun, 08 Jul 2018 07:08:23 +0000',
         // colors from http://www.w3.org/wiki/CSS/Properties/color/keywords
         color_names: [
             'transparent', 'currentcolor', 'black', 'silver', 'gray', 'white',
@@ -3078,6 +3058,79 @@
             return str.split(format_split_re).filter(Boolean);
         },
         // ---------------------------------------------------------------------
+        // :: replace that return position after replace for working with
+        // :: replacement that change length of the string
+        // :: source https://stackoverflow.com/a/46756077/387194
+        // ---------------------------------------------------------------------
+        tracking_replace: function tracking_replace(string, rex, replacement, position) {
+            function substring(string, start, end) {
+                return $.terminal.substring(string, start, end);
+            }
+            function length(string) {
+                return $.terminal.length(string);
+            }
+            var new_string = "";
+            var match;
+            var index = 0;
+            var rep_string;
+            var new_position = position;
+            var start;
+            var global = rex.flags.indexOf('g') !== -1;
+            rex.lastIndex = 0; // Just to be sure
+            while ((match = rex.exec(string))) {
+                // if regex don't have g flag lastIndex don't work
+                if (global) {
+                    // Add any of the original string we just skipped
+                    var last_index = length(substring(string, 0, rex.lastIndex));
+                    start = last_index - length(match[0]);
+                } else {
+                    start = match.index;
+                    last_index = start + length(match[0]);
+                }
+                if (index < start) {
+                    new_string += substring(string, index, start);
+                }
+                index = last_index;
+                // Build the replacement string. This just handles $$ and $n,
+                // you may want to add handling for $`, $', and $&.
+                rep_string = replacement.replace(/\$(\$|\d)/g, function(m, c0) {
+                    if (c0 === "$") {
+                        return "$";
+                    }
+                    return match[c0];
+                });
+                // Add on the replacement
+                new_string += rep_string;
+                // If the position is affected...
+                if (start < position) {
+                    // ... update it:
+                    if (last_index < position) {
+                        // It's after the replacement, move it
+                        new_position = Math.max(
+                            0,
+                            new_position +
+                            length(rep_string) -
+                            length(match[0])
+                        );
+                    } else {
+                        // It's *in* the replacement, put it just after
+                        new_position += length(rep_string) - (position - start);
+                    }
+                }
+                // If the regular expression doesn't have the g flag, break here so
+                // we do just one replacement (and so we don't have an endless loop!)
+                if (!global) {
+                    break;
+                }
+            }
+            // Add on any trailing text in the string
+            if (index < length(string)) {
+                new_string += substring(string, index);
+            }
+            // Return the string and the updated position
+            return [new_string, new_position];
+        },
+        // ---------------------------------------------------------------------
         // :: helper function used by substring and split_equal it loop over
         // :: string and execute callback with text count and other data
         // ---------------------------------------------------------------------
@@ -3100,7 +3153,8 @@
             }
             // ----------------------------------------------------------------
             function is_text(i) {
-                return not_formatting && string[i] !== ']' && !opening;
+                return not_formatting && (string[i] !== ']' || !have_formatting)
+                    && !opening;
             }
             // ----------------------------------------------------------------
             var have_formatting = $.terminal.have_formatting(string);
@@ -3415,8 +3469,10 @@
         // :: apply custom formatters only to text
         // ---------------------------------------------------------------------
         apply_formatters: function(string, settings) {
-            function test_lengths(ret, string) {
-                if ($.terminal.length(ret) !== $.terminal.length(string)) {
+            function test_lengths(formatter, ret, string) {
+                if (!formatter.__no_warn__ &&
+                    $.terminal.length(ret) !== $.terminal.length(string)) {
+                    console.log(formatter);
                     warn('Your formatter change length of the string, ' +
                          'you should use [regex, replacement] formatte' +
                          'r instead');
@@ -3433,7 +3489,7 @@
                     // on the list
                     if (typeof formatter === 'function' && formatter.__meta__) {
                         var ret = formatter(string, settings);
-                        test_lengths(ret, string);
+                        test_lengths(formatter, ret, string);
                         if (typeof ret === 'string') {
                             return ret;
                         }
@@ -3443,10 +3499,20 @@
                                 return string;
                             } else {
                                 if (formatter instanceof Array) {
+                                    var options = formatter[2] || {};
+                                    if (options.loop) {
+                                        while (string.match(formatter[0])) {
+                                            string = string.replace(
+                                                formatter[0],
+                                                formatter[1]
+                                            );
+                                        }
+                                        return string;
+                                    }
                                     return string.replace(formatter[0], formatter[1]);
                                 } else if (typeof formatter === 'function') {
                                     var ret = formatter(string, settings);
-                                    test_lengths(ret, string);
+                                    test_lengths(formatter, ret, string);
                                     if (typeof ret === 'string') {
                                         return ret;
                                     }
@@ -3460,7 +3526,7 @@
             } catch (e) {
                 var msg = 'Error in formatter [' + (i - 1) + ']';
                 formatters.splice(i - 1);
-                throw new $.terminal.Exception('formatting', msg);
+                throw new $.terminal.Exception('formatting', msg, e.stack);
             }
         },
         // ---------------------------------------------------------------------
@@ -3603,6 +3669,12 @@
         // ---------------------------------------------------------------------
         escape_brackets: function escape_brackets(string) {
             return string.replace(/\[/g, '&#91;').replace(/\]/g, '&#93;');
+        },
+        // ---------------------------------------------------------------------
+        // :: complmentary function
+        // ---------------------------------------------------------------------
+        unescape_brackets: function unescape_brackets(string) {
+            return string.replace(/&#91;/g, '[').replace(/&#93;/g, ']');
         },
         // ---------------------------------------------------------------------
         // :: return number of characters without formatting
@@ -3851,13 +3923,16 @@
         })()
     };
     // -------------------------------------------------------------------------
-    $.terminal.Exception = function Terminal_Exception(type, message) {
+    $.terminal.Exception = function Terminal_Exception(type, message, stack) {
         if (arguments.length === 1) {
             this.message = arguments[0];
             this.type = 'TERMINAL';
         } else {
             this.type = type;
             this.message = message;
+            if (stack) {
+                this.stack = stack;
+            }
         }
     };
     $.terminal.Exception.prototype = new Error();
@@ -3941,7 +4016,7 @@
                     if (options.error) {
                         options.error(jqXHR, 'Invalid JSON', e);
                     } else {
-                        throw new $.terminal.Exception('JSON', 'Invalid JSON');
+                        throw new $.terminal.Exception('JSON', 'Invalid JSON', e.stack);
                     }
                     deferred.reject({message: 'Invalid JSON', response: response});
                     return;
@@ -4106,6 +4181,8 @@
     // :: Default options
     // -----------------------------------------------------------------------
     $.terminal.nested_formatting.__meta__ = true;
+    // nested formatting will always return different length so we silent the warning
+    $.terminal.nested_formatting.__no_warn__ = true;
     $.terminal.defaults = {
         prompt: '> ',
         history: true,